Move DOM Manipulation to a separate task
Calculating a new point, it's position and painting the dot doens't need
to depend on the score/calculation boxes being updated.

This puts updating DOM element content into the requestAnimationFrame loop
which happens to make the calculation neater to read also.

By storing the loop timers this way we now get start and pause features
within the function.